+++ Models/ComprasDetalles.cs
@@ -9,7 +9,7 @@ public class ComprasDetalles
     public int DetalleId { get; set; }
     public int CompraId { get; set; }
     public int Cantidad { get; set; }
-    public decimal Precio { get; set; }
+    public float Precio { get; set; }
     public int ProductoId { get; set; }
 
     [ForeignKey("ProductoId")]
